fix: filter invalid categories in "etl_pipeline.py"
data extracted included invalid values for categories (strings which were not
ending with '0' or '1'. This data was on the end included in the database used for training a model.

To fix that bug a new function "lastchar_to_01" was included which checks not only
extracts the last character and transforms it to an integer, but also checks
the final value to be within [0, 1]. Else NaN is returned.

The function "transform_data" is adapted so that any data-points which include nan-values in
the categories will be dropped.